separate string similiar code into separate file
- back fill unit tests for this code
